analytics-service/requirements.txt
# Analytics Service Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.12
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Database
supabase==2.0.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Redis for caching
redis==5.0.1
aioredis==2.0.1

# HTTP clients
httpx==0.25.2
aiohttp==3.9.1

# Async support
asyncio-mqtt==0.16.1

# Data processing
python-dateutil==2.8.2
pytz==2023.3

# Development
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
